move restaurant logic to separate domain folder
