+++ apps/backend/src/restaurants/index.ts
@@ -1,0 +1,4 @@
+export { RestaurantService } from "./restaurant.service";
+export { Restaurant, RestaurantSearchParams, Review } from "./restaurant.types";
+
+export { default as mockRestaurantData } from "./restaurants-mock.json";
