+++ apps/backend/src/rooms/room.types.ts
@@ -1,4 +1,5 @@
 import { User } from "../types";
+import { Restaurant } from "../restaurants";
 
 // TODO: Add timer
 export type CreateRoomRequest = {
@@ -20,20 +21,6 @@ export type JoinRoomRequest = {
   userName: string;
 };
 
-// We use strings for monetary values bc 64 bit values are represented as strings in JSON
-export type Restaurant = {
-  id: string;
-  name: string;
-  address: string;
-  rating: number;
-  minPrice: string;
-  maxPrice: string;
-  mapLink: string | undefined | null;
-  openNow: boolean;
-  reviews: Review[];
-  photos: string[];
-};
-
 export type RoomResponse = {
   id: string;
   owner: User;
@@ -43,11 +30,6 @@ export type RoomResponse = {
   restaurants: Restaurant[];
   restaurantVotes: { [id: string]: number };
   gameState: GameState;
-};
-
-export type Review = {
-  rating: number;
-  text: string;
 };
 
 export type GameState = "LOBBY" | "STARTED" | "FINISHED";
